requirements.txt
anthropic
httpx
beautifulsoup4
pydantic
python-dotenv
streamlit
pytest
